requirements.txt
asgiref==3.6.0
Django==4.1.6
sqlparse==0.4.3
tzdata==2022.7
Celery

flake8
flake8-builtins
flake8-print

django-admin-rangefilter
django-crispy-forms
crispy-bootstrap5
django-filter
djangorestframework
djangorestframework-xml
djangorestframework-yaml
djangorestframework-simplejwt
drf-yasg